[MERGE #4077 @MikeHolman] Optimize JSON
Merge pull request #4077 from MikeHolman:lazyjson

This change represents a major overhaul to JSON.stringify.

The new architecture works as follows:

We first calculate the length of the string we need and create a side data structure with the values we will need, but do not actually create any concat strings, etc. like we used to do.

If/when the string is needed, we create a buffer using the length and metadata collected in the first pass, filling it element by element.

If the string is passed to JSON.parse, we use our metadata to construct an object without needing to parse the string.

By avoiding creation of all the intermediate concat strings, this makes JSON.stringify much more efficient. In micro-benchmarks I'm seeing JSON.stringify take around 40% less time than today.

And in case the string is actually sent to JSON.parse, that now takes 70% less time.

This improves Speedometer 2 geomean by about 3%. Also improves AcmeAir by about 3%.