[MERGE #4221 @ricobbe] Remove repeated function calls to improve OACR results
Merge pull request #4221 from ricobbe:build/ricobbe/bug-10437444-on-1.8

OS:10437444: Instead of calling functionBody->GetConstantCount() 5 times in JITTimeFunctionBody::InitializeJITFunctionData, call it only once and save the results in a local variable.  This allows OACR to conclude that the arithmetic and array index operation on line 53 are safe.  (There may also be a slight perf improvement, but this is incidental.)

We assume that all of the original calls to GetConstantCount() return the same value.  I've done a OneCirro run to validate this, and it turned up no errors.