[MERGE #3221 @xiaoyinl] Replace GetModuleFileName with GetBinaryLocation (fix #2297)
Merge pull request #3221 from xiaoyinl:replace_getmodulefilename

This will fix the remaining task in Issue #2297: replace `GetModuleFileName` with `PlatformAgnostic::SystemInfo::GetBinaryLocation`.

This task is easy if the caller of `GetModuleFileName` can easily work with a UTF-8 string, like `TTDHostBuildCurrentExeDirectory`. But if the caller needs to pass the returned string to a Unicode version of Windows API (see #3178 for example), then it seems the UTF-8 string returned by `GetBinaryLocation` is not so suitable here. What to do in this case?