[MERGE #4264 @aneeshdk] Updating the scope->GetIsObject() checks to fail fast
Merge pull request #4264 from aneeshdk:ScopeIsObjectAssert

The places where we assert that the scope is an object, we follow it up with a bytecode which treats the specified scope as an object. In error cases, when the scope is not an object, we will have type confusion in release bits. So changing such asserts to fail fast.