[MERGE #4293 @sigatrev] ARM64: ChangeToAssign with Indirs
Merge pull request #4293 from sigatrev:indir

assigns from IndirOpnds get converted to LDR which can handle 8/16 byte sources. Changing the type of the source of an LDR is can result in the LDR reading some memory off the end of the array, and while those bytes are discarded by the UBFX/SBFX it can result in an access violation if reading at a page boundary.