[MERGE #4349 @akroshg] Removing some marshalling code for bound function
Merge pull request #4349 from akroshg:boundmarshal

In order to fix OS14392550, I was trying to make the changes in the bound function. However with few testing it is determined that we don't need to  have this unnecessary marshalling code.
Added few test-cases to see if the marshalling does work as expected.