[MERGE #4384 @aneeshdk] Adding an option to provide a timeout value through command line in RL
Merge pull request #4384 from aneeshdk:RLTimeout

When tests run on lower configuration machines the timeout values specified in the XML does not seem to be valid. This change is to provide a way to specify a timeout value for all tests cases in such cases. This will override the value from the XML file.