[PR #4381 @digitalinfinity] Devirtualize FunctionInfo
Update the core submodule and also update HeapEnum to not depend on FunctionInfo::Is

[Core Commit Message]

[MERGE #4381 @digitalinfinity] Devirtualize FunctionInfo

Merge pull request #4381 from digitalinfinity:devirt_funcinfo

With the changes to the FunctionBody hierarchy for redeferral, FunctionInfo
no longer needs to inherit from FinalizableObject. Removing the base class
gets rid of the vtable pointer and shrinks it by 8 bytes and into the 32 byte
bucket in the Recycler. This saves ~500k on Gmail.