[MERGE #4104 @atulkatti] Enable the allocations during concurrent sweep - Part2.
Merge pull request #4104 from atulkatti:ConcurrentSweepAllocations.Part2.master.1

If allocations are to be allowed to existing heap blocks during concurrent sweep, we set aside a few
heap blocks from the heapBlockList prior to beginning sweep. However, we eed to then go back and make sure these blocks also swept before this sweep finishes. In order to do this we clearly define concurrent sweep having 2 passes now. These passes existed before but were not distiguished as they would always start and finish in one go on the background thread. However, whenever allocations are allowed during concurrent sweep; the concurrent sweep will start Pass1 on the background thread, wait to finish Pass1 of the blocks we set aside to allocate from on the main thread and then go back to finish Pass2 for all heap blocks on the background thread. Note that, due to this need to finish Pass1 on the foreground thread the overall background sweep will now appear to take longer whenever we chose to do such a two-pass sweep.

The sequence of things we do to allow allocations during concurrent sweep is described below:
1. At the beginning of concurrrent sweep we decide if we will benefit from allowing allocations during concurrent sweep for any of the buckets. If there is at-least one bucket for which we think we will benefit we will turn on allocations during concurrent sweep. Once turned on we will attempt to enable allocations during concurrent sweep for all supported buckets (i.e. small/medium, normal/leaf, non-finalizable buckets.write barrrier bickets are supported as well.).
2. If allocations are turned on during concurrent sweep, we will see if there are any allocable blocks in the
heapBlockList after the nextAllocableBlockHead. If we find any such blocks, we move them to a SLIST that the allocator can pick these blocks from during sweep.
3. CollectionStateConcurrentSweepPass1: We will finish Pass1 of the sweep for all the remaining blocks (other than the ones we put in the SLIST in step 2 above) This will generally happen on the background thread unless we are forcing in-thread sweep. This state is now specifically identified as CollectionStateConcurrentSweepPass1.
4. CollectionStateConcurrentSweepPass1Wait: At this point we need to wait for all the blocks that we put in the SLIST to also finish the Pass1 of the sweep. This needs to happen on the foreground thread so we prevent the allocator from picking up the blocks from SLIST while we do this. This state is now identified as CollectionStateConcurrentSweepPass1Wait.
5. CollectionStateConcurrentSweepPass2: At this point we will do the actual sweeping of all the blocks that are not yet swept, for eaxample, any blocks that were put onto the pendingSweepList. As these blocks get swept we keep adding them to the SLIST again to allow allocators to allocate from them as soon as they are swept.
6. Before the Pass2 can finish we can call this concurrent sweep done we need to move all the blocks off of the SLIST so that normal allocations can begin after the sweep. This is the last step of the concurrent sweep.