[MERGE #4419 @Penguinwizzard] A pass over initialization of Arena-allocated classes
Merge pull request #4419 from Penguinwizzard:initializer_reworks

PREfast raised warnings on a number of classes about members lacking
initialization before the constructor, and in many cases after it. A
comparison between that list and the list of classes that we make in
arena memory (where we often don't zero) gave the set addressed here
in the associated files.