[MERGE #4458 @thomasmo] Uint8ClampedArray,Int32Array.set does not clamp correctly when virtual
Merge pull request #4458 from thomasmo:virtualarray_set

Some TypedArrays can become virtual when they meet a set of requirements (defined in JavascriptArrayBuffer::IsValidAsmJsBufferLengthAlgo). These TypedArrays become a separate type with a separate VTable at runtime. Because of this difference, virtual array checks are missing in TypedArrayBase::Set for the fast path of setting the buffer, resulting in numerical type mismatch when copying. This change fixes the check to also account for virtual TypedArrays.
An additional force check is added to make the host always allocate TypedArrays as virtual. A few tests are updated with this flag to prevent future regressions.