[MERGE #4652 @ricobbe] Use uint16 for the number of capturing groups in Regex interfaces
Merge pull request #4652 from ricobbe:build/ricobbe/regex-numgroups-uint16-on-master

In RegexParser, RegexPattern, and RegexRuntime, change the type of the number of capturing groups from int to uint16 (since we limit the number of groups to 2^15-1 in the parser).

Follow-on to MSRC 41416), which addressed a potential arithmetic overflow and subsequent illegal array access.