[MERGE #4721 @ricobbe] OS#15992535 Revert #4652 to fix potential uint16 overflow in RegexParser
Merge pull request #4721 from ricobbe:build/ricobbe/bug-15992535

My commit in PR #4652 introduced the possibility of an overflow in the regex parser: by changing RegexParser::numGroups and RegexParser::nextGroupId from int to uint16, we make it possible to overflow nextGroupId before hitting the check for an excessive number of capturing groups.

This PR is to remove this incorrect behavior before it propagates out of DEV3; I'll follow up with a later commit to adjust this API in a safer manner (and also possibly replace the AssertOrFailFast in RegexParser with something a little bit more user-friendly).