[MERGE #4754 @jackhorton] OS#16115338 Update _.floor to normalize -0 to 0 [OSS-Fuzz]
Merge pull request #4754 from jackhorton:os16115338

The intl spec refers to the `floor` function in [DefaultNumberOption](https://tc39.github.io/ecma402/#sec-defaultnumberoption), which we had previously implemented as exactly Math.floor Math.floor is defined to return -0 when given -0 as input. However, DefaultNumberOption actually refers to [the floor mathematical operation](https://tc39.github.io/ecma262/#sec-mathematical-operations), which states that -0 should be normalized to 0, which is the more intuitive behavior.

This fixes a FailFast that triggered when we couldn't retrieve the minimumFractionDigits as a TaggedInt.