[MERGE #4776 @akroshg]  Fix for re-entrancy in splice and unshift
Merge pull request #4776 from akroshg:splice

When we have getter/setter at 4294967295, we go out in the user code and we end up throwing the re-entrancy error.
For splice we now take the slow path for the overflow case.
for unshift - just put the re-entrant macro around it and also handle the es5 array case.