[MERGE #4804 @irinayat-MS] OS#16244108: Small refactor of Parser::CreateCallNode and Parser::CreateSuperCallNode
Merge pull request #4804 from irinayat-MS:ParseNodeCall

https://microsoft.visualstudio.com/OS/_workitems/edit/16244108

The uninitialized field was introduced by https://github.com/Microsoft/ChakraCore/pull/3917. The only read of the field is in EmitArgList, and if it ends up "true" instead of the default "false" an extra defensive load will be emitted for the constructor parameters so not a security/correctness concern.