[MERGE #4777 @benjamin-bader] Add ccache option to CMakeLists.txt and build.sh
Merge pull request #4777 from benjamin-bader:bebader/ccache

This PR adds two new, related, flags to build.sh, `--use-ccache` and `--ccache=NAME`.  The former tells cmake to use `ccache` when building ChakraCore, which will greatly speed up "build-the-world" events both for local devs and for CI builds.  In our particular case (building ChakraCore as a static library as part of our product's build), we've reduced the Chakra build time from 25 minutes to 1 minute on average.

The latter flag implies the former, and allows one to specify a custom ccache-compatible program.  This is useful for when using something like https://github.com/mozilla/sccache/, which uses cloud storage as a cache and supports MSVC.

This is a massive win for our CI; hopefully, with sccache, it can be a win for yours, too.

Fixes #4775