[MERGE #4861 @pleath] [MSFT 16510487] Change type when assigning cross site entry point to a locked type
Merge pull request #4861 from pleath:16510487

We've been evolving shared types (only) when we set their entry points to a cross-site thunk. Now that function types can belong to type paths, we need to do this when types are locked as well. This avoids the case where a function gets a cross-site type when a property is added to the object.