[MERGE #4869 @pleath] [MSFT 16526028] Don't let cross-site functions have PathTypeHandler's
Merge pull request #4869 from pleath:16526028

Naively letting a function type participate in a type path when its entryPoint is set to a cross-site thunk means that it can transition to a non-cross-site type by having properties added or attributes set on existing properties. If a function type is already cross-site on first property access, go to a non-shared SimpleTypeHandler instead of PathTypeHandler. If a function type with a PathTypeHandler becomes a cross-site type, go to (Simple)DictionaryTypeHandler.