[MERGE #4907 @sethbrenith] Cache property enumeration in JSON.stringify like we do for Object.assign
Merge pull request #4907 from sethbrenith:user/sethb/stringify-cache-iter

JSON payloads tend to be relatively repetitive, so we can save a lot on large serializations by reusing object property iteration. On a microbenchmark that serializes and parses a repetitive piece of data that I found in Speedometer, I see a 35% improvement.