[MERGE #4908 @rhuanjl] Class extension of proxy without construct trap fix #4729
Merge pull request #4908 from rhuanjl:proxySuper

Proxy trap function when there's no construct trap was always making a new object as the "this" to use - in the case of a super call this would break the class inheritance chain.

Currently CC's behaviour on this differers from jsc, v8 and sm. This fix brings CC into line with the others.

Should fix #4729