[MERGE #4985 @digitalinfinity] Fix build break
Merge pull request #4985 from digitalinfinity:build_break

The recent ETW manifest change assumes that UInt8 can be represented by the output type xs:boolean. However, this is a change supported only by mc.exe version 10.0.14251. In versions older than that, it breaks (which is also why github CI didn't catch it). Switched to using the Win32 BOOL input type instead.

There is a second issue where recycler sweep's pointer to the recycler might not be valid at the point of the pre-sweep callback. Updated to account for this