[MERGE #4981 @rhuanjl] Dataview set/get negative index throw RangeError Fixes: #4978
Merge pull request #4981 from rhuanjl:dataview

Simple fix for #4978

Note the ToIndex method I used was based on the implementation in ArrayBuffer.cpp but slightly simpler as DataView.cpp has a check for lengths that are too big later on.

Could clean up the code slightly by using a consistent version of ToIndex in both places though that would be a more invasive change than this.