[MERGE #5058 @thomasmo] Initial commit for BGParseManager
Merge pull request #5058 from thomasmo:bgparse_intro

Initial commit for BGParseManager.
BGParseManager is responsible for handling and managing requests to parse scripts on a background thread. This work is tracked by the BGParseWorkItem with a cookie, and is managed by BackgroundJobProcessor. The results of the background work is marshaled across threads through bytecode serialization.More details in the header file for BGParseManager.

This work can be controlled by a new config flag, BgParse; additional tracing at runtime is also managed by the same flag;