[MERGE #5077 @sethbrenith] Let substrings get property string pointers too
Merge pull request #5077 from sethbrenith:user/sethb/sub-prop-string-master

We currently have a system to promote CompoundString or ConcatString instances to become LiteralStringWithPropertyStringPointer. This applies the same transformation to SubString. Also:

- Call GetString instead of GetSz in ThreadContext::FindPropertyRecord to avoid an extra copy
- Actively cache the result in ThreadContext::FindPropertyRecord to avoid a separate lookup later
- Update the buffer in LiteralStringWithPropertyStringPtr to point to the copy of the string allocated alongside the PropertyRecord, which could free up some memory in some cases

With this change, I see about 8% speedup on the Ember test in Speedometer, because Ember uses many strings that begin their lifetime as substrings.

We are still emitting a helper call to OP_GetElementI for the particular function I was focusing on improving (in its entirety, `function(e){return this.data[e]}`) because our profile data happens to be all cache misses, so we have no indication that we would eventually get a bunch of local-cache hits. At least we can now use the fast path in the helper, which seems to make a big difference.