[MERGE #4995 @rajatd] Aggressive Value Transfer in loop prepass
Merge pull request #4995 from rajatd:preInline

This change enables us to transfer values in the prepass of a loop. This enables us to do more aggressive object pointer copy-prop in the prepass, which is a precursor to getting field PRE kick in in more cases.

The basic condition that needs to be satisfied to transfer value from `src` to `dst` in the pre-pass is that either the `src` shouldn't be live on back edge of the current loop or, if it is, it shouldn't be written to in the loop (or any of the parent loops).

I have also included a change that will enable field PRE in cases when the `this` object is passed to an inlinee that loads properties from that object. Those loads could now be PRE'd.