[MERGE #5124 @Cellule] MemCopy ensure src head
Merge pull request #5124 from Cellule:memcpy

Ensure we always use the head segment of the source array in MemCopy optimization.
Currently the bounds check condition needed in order to have a memcpy ensures that we use the head segment. But in the memcpy code it is not obvious so make it explicit.